Move SpendingInfoDb to core wallet api
