+++ app/server-test/src/test/scala/org/bitcoins/server/RoutesSpec.scala
@@ -8,6 +8,7 @@ import akka.http.scaladsl.testkit.ScalatestRouteTest
 import org.bitcoins.chain.api.ChainApi
 import org.bitcoins.commons.jsonmodels.wallet.CoinSelectionAlgo
 import org.bitcoins.core.Core
+import org.bitcoins.core.api.wallet.db.{SegwitV0SpendingInfo, SpendingInfoDb}
 import org.bitcoins.core.config.RegTest
 import org.bitcoins.core.crypto.ExtPublicKey
 import org.bitcoins.core.currency.{Bitcoins, CurrencyUnit, Satoshis}
@@ -29,8 +30,6 @@ import org.bitcoins.core.wallet.utxo.{
   AddressLabelTagType,
   AddressTag,
   AddressTagType,
-  SegwitV0SpendingInfo,
-  SpendingInfoDb,
   TxoState
 }
 import org.bitcoins.crypto.{
