+++ wallet-test/src/test/scala/org/bitcoins/wallet/WalletCallbackTest.scala
@@ -1,8 +1,8 @@
 package org.bitcoins.wallet
 
+import org.bitcoins.core.api.wallet.db.SpendingInfoDb
 import org.bitcoins.core.protocol.BitcoinAddress
 import org.bitcoins.core.protocol.transaction.{EmptyTransaction, Transaction}
-import org.bitcoins.core.wallet.utxo.SpendingInfoDb
 import org.bitcoins.testkit.wallet.BitcoinSWalletTest
 import org.bitcoins.testkit.wallet.FundWalletUtil.FundedWallet
 import org.scalatest.FutureOutcome
