+++ wallet/src/main/scala/org/bitcoins/wallet/api/HDWalletApi.scala
@@ -1,6 +1,7 @@
 package org.bitcoins.wallet.api
 
 import org.bitcoins.commons.jsonmodels.wallet.CoinSelectionAlgo
+import org.bitcoins.core.api.wallet.db.SpendingInfoDb
 import org.bitcoins.core.currency.CurrencyUnit
 import org.bitcoins.core.hd.{AddressType, HDAccount, HDChainType, HDPurpose}
 import org.bitcoins.core.protocol.BitcoinAddress
